# api/openapi/spec.yaml
openapi: "3.0.3"
info:
  title: Kill-AI-Leak API
  description: |
    REST API for the Kill-AI-Leak AI security platform.

    Kill-AI-Leak provides real-time guardrails, inventory management, policy
    enforcement, and event auditing for AI/LLM workloads across your
    organization. Every request must include a valid bearer token and an
    application identifier header.
  version: "1.0.0"
  contact:
    name: Kill-AI-Leak Maintainers
    url: https://github.com/kill-ai-leak/kill-ai-leak
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.kill-ai-leak.local
    description: Default cluster-local endpoint
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
    AppID: []
...
paths:
...
  /api/v1/events:
    get:
      operationId: listEvents
      summary: Query security events
      description: |
        Returns a paginated, filterable list of security events. Supports
        filtering by source, severity, actor, time range, and more.
      tags: [events]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/SortByParam"
        - $ref: "#/components/parameters/SortOrderParam"
        - name: source
          in: query
          description: Filter by event source.
          schema:
            type: string
            enum:
              - kernel_observer
              - inline_gateway
              - browser
              - ide
              - mcp_gateway
              - cicd
        - name: severity
          in: query
          description: Filter by minimum severity level.
          schema:
            type: string
            enum: [info, low, medium, high, critical]
        - name: actor_id
          in: query
          description: Filter by actor ID.
          schema:
            type: string
        - name: actor_type
          in: query
          description: Filter by actor type.
          schema:
            type: string
            enum: [pod, user, service_account, browser_user, agent]
        - name: provider
          in: query
          description: Filter by target LLM provider.
          schema:
            type: string
        - name: decision
          in: query
          description: Filter by guardrail decision.
          schema:
            type: string
            enum: [allow, block, anonymize, modify, alert, coach, throttle, log]
        - name: blocked
          in: query
          description: Filter by whether the request was blocked.
          schema:
            type: boolean
        - name: session_id
          in: query
          description: Filter by session ID.
          schema:
            type: string
        - name: start_time
          in: query
          description: Return events after this timestamp (RFC 3339).
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          description: Return events before this timestamp (RFC 3339).
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated list of events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/policies:
    get:
      operationId: listPolicies
      summary: List all policies
      description: Returns all configured AI security policies.
      tags: [policies]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/SortByParam"
        - $ref: "#/components/parameters/SortOrderParam"
        - name: namespace
          in: query
          description: Filter by policy namespace.
          schema:
            type: string
        - name: mode
          in: query
          description: Filter by enforcement mode.
          schema:
            type: string
            enum: [off, discover, monitor, enforce]
      responses:
        "200":
          description: Paginated list of policies.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createPolicy
      summary: Create a new policy
      description: Creates a new AI security policy.
      tags: [policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AISecurityPolicy"
            example:
              apiVersion: kill-ai-leak.io/v1
              kind: AISecurityPolicy
              metadata:
                name: production-default
                namespace: production
              spec:
                scope:
                  namespaces: [production]
                providers:
                  allow: [openai, anthropic]
                  deny: []
                input:
                  block_pii: false
                  anonymize_pii: true
                  pii_types: [ssn, credit_card]
                  block_secrets: true
                  block_injection_score_above: 0.85
                mode: enforce
      responses:
        "201":
          description: Policy created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AISecurityPolicy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: A policy with this name already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token for authentication.
    AppID:
      type: apiKey
      in: header
      name: X-App-ID
      description: Application identifier header. Ties requests to a registered application.

  # ---------- reusable parameters ----------
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    SortByParam:
      name: sort_by
      in: query
      description: Field to sort by.
      schema:
        type: string
        default: created_at
    SortOrderParam:
      name: sort_order
      in: query
      description: Sort direction.
      schema:
        type: string
        enum: [asc, desc]
        default: desc

  # ---------- reusable responses ----------
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code.
          example: "invalid_request"
        message:
          type: string
          description: Human-readable error description.
          example: "prompt field is required"
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: JSON path of the offending field.
          example: "prompt"
        reason:
          type: string
          description: Why this field failed validation.
          example: "must not be empty"
    Pagination:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 50
        total_items:
          type: integer
          example: 234
        total_pages:
          type: integer
          example: 5

    # -- Actor / Target / Action --
    Actor:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
          enum: [pod, user, service_account, browser_user, agent]
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        node:
          type: string
        service_account:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        team:
          type: string
    Target:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
          enum: [llm_provider, mcp_server, database, filesystem, api]
        id:
          type: string
        provider:
          type: string
        endpoint:
          type: string
        model:
          type: string
        region:
          type: string
    Action:
      type: object
      required: [type, direction]
      properties:
        type:
          type: string
          enum: [api_call, tool_exec, file_access, process_spawn, db_query]
        direction:
          type: string
          enum: [outbound, inbound]
        protocol:
          type: string
        method:
          type: string

    # -- Content / Guardrail results --
    ContentMeta:
      type: object
      required: [has_prompt, blocked, anonymized]
      properties:
        has_prompt:
          type: boolean
        prompt_hash:
          type: string
        prompt_text:
          type: string
        response_hash:
          type: string
        response_text:
          type: string
        tokens_input:
          type: integer
        tokens_output:
          type: integer
        model:
          type: string
        pii_detected:
          type: array
          items:
            type: string
        injection_score:
          type: number
          format: double
        blocked:
          type: boolean
        anonymized:
          type: boolean
    GuardrailResult:
      type: object
      required: [rule_id, rule_name, stage, decision, confidence, latency_ms]
      properties:
        rule_id:
          type: string
        rule_name:
          type: string
        stage:
          type: string
          enum: [pre_input, input, routing, output, post_output, behavioral]
        decision:
          type: string
          enum: [allow, block, anonymize, modify, alert, coach, throttle, log]
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        reason:
          type: string
        details:
          type: array
          items:
            type: string
        latency_ms:
          type: integer
          format: int64
...
    Event:
      type: object
      required: [id, timestamp, source, severity, actor, target, action, content]
      properties:
        id:
          type: string
          description: ULID event identifier.
          example: "01HZ3XJKB0001ABCDEF"
        timestamp:
          type: string
          format: date-time
        source:
          type: string
          enum: [kernel_observer, inline_gateway, browser, ide, mcp_gateway, cicd]
        severity:
          type: string
          enum: [info, low, medium, high, critical]
        actor:
          $ref: "#/components/schemas/Actor"
        target:
          $ref: "#/components/schemas/Target"
        action:
          $ref: "#/components/schemas/Action"
        content:
          $ref: "#/components/schemas/ContentMeta"
        guardrails:
          type: array
          items:
            $ref: "#/components/schemas/GuardrailResult"
        metadata:
          type: object
          additionalProperties:
            type: string
        session_id:
          type: string
        trace_id:
          type: string
        cost_usd:
          type: number
          format: double
        latency_ms:
          type: integer
          format: int64
    EventListResponse:
      type: object
      required: [events, pagination]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # -- Inventory --
...
    PolicyScope:
      type: object
      properties:
        namespaces:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        service_accounts:
          type: array
          items:
            type: string
        users:
          type: array
          items:
            type: string
        teams:
          type: array
          items:
            type: string
    ProviderPolicy:
      type: object
      required: [allow, deny]
      properties:
        allow:
          type: array
          items:
            type: string
        deny:
          type: array
          items:
            type: string
        namespace_overrides:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProviderPolicy"
    ModelPolicy:
      type: object
      required: [allow, deny]
      properties:
        allow:
          type: array
          items:
            type: string
        deny:
          type: array
          items:
            type: string
    RateLimit:
      type: object
      properties:
        requests_per_minute:
          type: integer
        requests_per_hour:
          type: integer
        requests_per_day:
          type: integer
        tokens_per_day:
          type: integer
          format: int64
        cost_per_day_usd:
          type: number
          format: double
        cost_per_month_usd:
          type: number
          format: double
    RateLimitPolicy:
      type: object
      properties:
        per_user:
          $ref: "#/components/schemas/RateLimit"
        per_service:
          $ref: "#/components/schemas/RateLimit"
        per_namespace:
          $ref: "#/components/schemas/RateLimit"
    InputPolicy:
      type: object
      properties:
        block_pii:
          type: boolean
        anonymize_pii:
          type: boolean
        pii_types:
          type: array
          items:
            type: string
        block_secrets:
          type: boolean
        block_injection_score_above:
          type: number
          format: double
        max_tokens_per_request:
          type: integer
        blocked_topics:
          type: array
          items:
            type: string
        allowed_topics:
          type: array
          items:
            type: string
    OutputPolicy:
      type: object
      properties:
        block_toxic_score_above:
          type: number
          format: double
        scan_generated_code:
          type: boolean
        block_vulnerable_code:
          type: boolean
        check_pii_leakage:
          type: boolean
        check_prompt_leakage:
          type: boolean
    FilesystemPolicy:
      type: object
      required: [allow_read, deny_read, allow_write, deny_write, deny_delete]
      properties:
        allow_read:
          type: array
          items:
            type: string
        deny_read:
          type: array
          items:
            type: string
        allow_write:
          type: array
          items:
            type: string
        deny_write:
          type: array
          items:
            type: string
        deny_delete:
          type: array
          items:
            type: string
    CommandPolicy:
      type: object
      required: [allow, deny, require_approval]
      properties:
        allow:
          type: array
          items:
            type: string
        deny:
          type: array
          items:
            type: string
        require_approval:
          type: array
          items:
            type: string
    NetworkPolicy:
      type: object
      required: [allow_outbound, deny_outbound]
      properties:
        allow_outbound:
          type: array
          items:
            type: string
        deny_outbound:
          type: array
          items:
            type: string
    DatabasePolicy:
      type: object
      required: [allow, deny, require_approval]
      properties:
        allow:
          type: array
          items:
            type: string
        deny:
          type: array
          items:
            type: string
        require_approval:
          type: array
          items:
            type: string
    AgentPolicy:
      type: object
      properties:
        filesystem:
          $ref: "#/components/schemas/FilesystemPolicy"
        commands:
          $ref: "#/components/schemas/CommandPolicy"
        network:
          $ref: "#/components/schemas/NetworkPolicy"
        database:
          $ref: "#/components/schemas/DatabasePolicy"
    RouteTarget:
      type: object
      required: [provider]
      properties:
        provider:
          type: string
        region:
          type: string
        endpoint:
          type: string
    DataResidencyRule:
      type: object
      required: [name, condition, route_to]
      properties:
        name:
          type: string
        condition:
          type: string
        route_to:
          type: array
          items:
            $ref: "#/components/schemas/RouteTarget"
        deny_routing_to:
          type: array
          items:
            type: string
    DataResidencyPolicy:
      type: object
      required: [rules]
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/DataResidencyRule"
    AlertConfig:
      type: object
      properties:
        slack:
          type: string
        pagerduty:
          type: string
        email:
          type: string
        webhook:
          type: string
    PolicySpec:
      type: object
      required: [scope, mode]
      properties:
        scope:
          $ref: "#/components/schemas/PolicyScope"
        providers:
          $ref: "#/components/schemas/ProviderPolicy"
        models:
          $ref: "#/components/schemas/ModelPolicy"
        rate_limits:
          $ref: "#/components/schemas/RateLimitPolicy"
        input:
          $ref: "#/components/schemas/InputPolicy"
        output:
          $ref: "#/components/schemas/OutputPolicy"
        agent:
          $ref: "#/components/schemas/AgentPolicy"
        data_residency:
          $ref: "#/components/schemas/DataResidencyPolicy"
        mode:
          type: string
          enum: [off, discover, monitor, enforce]
        alerts:
          $ref: "#/components/schemas/AlertConfig"
    PolicyMetadata:
      type: object
      required: [name]
      properties:
        name:
          type: string
        namespace:
          type: string
    AISecurityPolicy:
      type: object
      required: [apiVersion, kind, metadata, spec]
      properties:
        apiVersion:
          type: string
          example: "kill-ai-leak.io/v1"
        kind:
          type: string
          example: AISecurityPolicy
        metadata:
          $ref: "#/components/schemas/PolicyMetadata"
        spec:
          $ref: "#/components/schemas/PolicySpec"
    PolicyListResponse:
      type: object
      required: [policies, pagination]
      properties:
        policies:
          type: array
          items:
            $ref: "#/components/schemas/AISecurityPolicy"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # -- Guardrail config --
...